itemsMapper function moved to links.util, UrlModel used in cms.model
