+++ src/galaxis-footer/contact-us/contact-us.component.tsx
@@ -1,50 +1,21 @@
-import React, { FC, useCallback, useEffect, useState } from "react";
+import React, { FC, useEffect, useState } from "react";
 import StyledWrapper from "./constact-us.style";
 import { Box, Link, Typography } from "@mui/material";
 import AddressModel from "../../models/address.model";
 import CustomTitle from "../title/title.component";
-import {
-  areUrlsSame,
-  extractSubjectFromUrl,
-  generateContactFormUrl,
-  getCurrentDomain,
-  removeDomainFromUrl,
-} from "../../utils/links.util";
+import { getCurrentDomain, useItemsMapper } from "../../utils/links.util";
 import { UrlModel } from "../../models/url.model";
 
 type Props = { hostName: string; socialMediaIcons: UrlModel[]; address: AddressModel };
 
 const ContactUs: FC<Props> = ({ hostName, socialMediaIcons, address }) => {
+  const itemsMapper = useItemsMapper();
+
   const [iconItems, setIconItems] = useState<UrlModel[]>(socialMediaIcons);
-
-  const iconsMapper = useCallback(
-    (currentHostName: string): UrlModel[] => {
-      return socialMediaIcons.map((item) => {
-        const currentUrl = getCurrentDomain(item.url);
-        const sameHost = areUrlsSame(currentHostName, currentUrl);
-
-        const subjectForContactUs = "custom subject here";
-        const generatedUrl = generateContactFormUrl(currentUrl, subjectForContactUs);
-
-        if (!sameHost) {
-          const path = generatedUrl ? generatedUrl : item.url;
-          return { ...item, url: path, openInNewTab: true };
-        } else {
-          let generatedPath = null;
-          if (generatedUrl) {
-            generatedPath = extractSubjectFromUrl(generatedUrl);
-          }
-          const path = removeDomainFromUrl(generatedPath ? generatedPath : item.url);
-          return { ...item, url: path, openInNewTab: false };
-        }
-      });
-    },
-    [socialMediaIcons]
-  );
 
   useEffect(() => {
     const currentHostName = getCurrentDomain(hostName);
-    const updatedPopular = iconsMapper(currentHostName);
+    const updatedPopular = itemsMapper(currentHostName, socialMediaIcons);
 
     setIconItems(updatedPopular);
   }, []);
