+++ src/galaxis-footer/pages/pages.component.tsx
@@ -1,35 +1,20 @@
-import React, { FC, useCallback, useEffect, useState } from "react";
+import React, { FC, useEffect, useState } from "react";
 import { Box, Link, Typography } from "@mui/material";
 import CustomTitle from "../title/title.component";
 import StyledWrapper from "./pages.style";
-import { areUrlsSame, getCurrentDomain, removeDomainFromUrl } from "../../utils/links.util";
+import { getCurrentDomain, useItemsMapper } from "../../utils/links.util";
 import { UrlModel } from "../../models/url.model";
 
 type PagesProps = { hostName: string; pages: UrlModel[] };
 
 const Pages: FC<PagesProps> = ({ hostName, pages }) => {
+  const itemsMapper = useItemsMapper();
+
   const [pagesItems, setPagesItems] = useState<UrlModel[]>(pages);
-
-  const pagesItemsMapper = useCallback(
-    (currentHostName: string): UrlModel[] => {
-      return pages.map((item) => {
-        const currentUrl = getCurrentDomain(item.url);
-        const sameHost = areUrlsSame(currentHostName, currentUrl);
-
-        if (!sameHost) {
-          return { ...item, openInNewTab: true };
-        } else {
-          const path = removeDomainFromUrl(item.url);
-          return { ...item, url: path, openInNewTab: false };
-        }
-      });
-    },
-    [pages]
-  );
 
   useEffect(() => {
     const currentHostName = getCurrentDomain(hostName);
-    const updatedPopular = pagesItemsMapper(currentHostName);
+    const updatedPopular = itemsMapper(currentHostName, pages);
 
     setPagesItems(updatedPopular);
   }, []);
