+++ src/models/cms.model.ts
@@ -1,22 +1,11 @@
 import AddressModel from "./address.model";
+import { UrlModel } from "./url.model";
 // import { UrlModel } from "./url.model";
 
 export interface CmsModel {
   aboutUs: string;
-  pages: {
-    label: string;
-    url: string;
-    openInNewTab: boolean;
-  }[];
-  socialMediaIcons: {
-    icon: string;
-    url: string;
-    openInNewTab: boolean;
-  }[];
+  pages: UrlModel[];
+  socialMediaIcons: UrlModel[];
   address: AddressModel;
-  popular: {
-    label: string;
-    url: string;
-    openInNewTab: boolean;
-  }[];
+  popular: UrlModel[];
 }
