+++ src/utils/links.util.ts
@@ -1,3 +1,6 @@
+import { useCallback } from "react";
+import { UrlModel } from "../models/url.model";
+
 const getCurrentDomain = (url: string): string => {
   const domainRegex = /^(?:https?:\/\/)?(?:dev\.|staging\.)?(.*?)\//;
   const matches = url.match(domainRegex);
@@ -47,10 +50,35 @@ const extractSubjectFromUrl = (url: string): string | null => {
   return null;
 };
 
+const useItemsMapper = () => {
+  return useCallback((currentHostName: string, items: UrlModel[]): UrlModel[] => {
+    return items.map((item) => {
+      const currentUrl = getCurrentDomain(item.url);
+      const sameHost = areUrlsSame(currentHostName, currentUrl);
+
+      const subjectForContactUs = "custom subject here";
+      const generatedUrl = generateContactFormUrl(currentUrl, subjectForContactUs);
+
+      if (!sameHost) {
+        const path = generatedUrl ? generatedUrl : item.url;
+        return { ...item, url: path, openInNewTab: true };
+      } else {
+        let generatedPath = null;
+        if (generatedUrl) {
+          generatedPath = extractSubjectFromUrl(generatedUrl);
+        }
+        const path = removeDomainFromUrl(generatedPath ? generatedPath : item.url);
+        return { ...item, url: path, openInNewTab: false };
+      }
+    });
+  }, []);
+};
+
 export {
   getCurrentDomain,
   areUrlsSame,
   removeDomainFromUrl,
   generateContactFormUrl,
   extractSubjectFromUrl,
+  useItemsMapper,
 };
